Redstone Furnace Tweaks
I am getting real tired of this furnace.  The furnace can pull items in
from the environment and a chest on top correctly. But when it comes to
putting an item into a chest it just doesn't wan to work. Think you can
give it a look to see if you can fix it?