Lot's and lot's of random stuff
So I tried to fix the mob rendering and stuff, but that didn't work... I
suggest that we just completely start over the mob, and work again from
ground zero.

Did some stuffs to the Netherrack Furnace

Posted some thingys on the wishlist AND the TODO list. Please tell me
what you think ;)