Formatting
I used the Eclipse shortcut "Ctrl+Shift+f" to auto-format the code. I'll
admit, it takes away a little readability, but it compresses the code
and makes it easier to view on smaller screen (which is helpful to me
since I have a big screen and small screen). It's also consistent in all
the code, which is something that always bugs me...